+++ async_job.go
@@ -1,4 +1,4 @@
-package async
+package gwf
 
 import (
 	"encoding/json"