+++ generate_key.go
@@ -1,4 +1,4 @@
-package command
+package gwf
 
 import (
 	"crypto/rand"
@@ -6,43 +6,40 @@ import (
 	"fmt"
 	"io/ioutil"
 	"strings"
-
-	gwf "github.com/RobyFerro/go-web-framework"
 )
 
 // GenerateKey will generate Go-Web application key in main config.yml file
 type GenerateKey struct {
 	Signature   string
 	Description string
+	Args        string
 }
 
 // Register this command
 func (c *GenerateKey) Register() {
-	c.Signature = "generate:key"             // Change command signature
-	c.Description = "Genate application key" // Change command description
+	c.Signature = "generate:key"               // Change command signature
+	c.Description = "Generate application key" // Change command description
 }
 
 // Run this command
-func (c *GenerateKey) Run(kernel *gwf.HttpKernel, args string, console map[string]interface{}) {
-	path := gwf.GetDynamicPath("config.yml")
+func (c *GenerateKey) Run() {
+	path := GetDynamicPath("config.yml")
 	read, err := ioutil.ReadFile(path)
 
 	if err != nil {
-		gwf.ProcessError(err)
+		ProcessError(err)
 	}
 
 	appKey, err := generateNewToken()
 	if err != nil {
-		gwf.ProcessError(err)
+		ProcessError(err)
 	}
 
 	newContent := strings.Replace(string(read), "$$APP_KEY$$", appKey, -1)
 
 	if err = ioutil.WriteFile(path, []byte(newContent), 0); err != nil {
-		gwf.ProcessError(err)
+		ProcessError(err)
 	}
-
-	return
 }
 
 // generateNewToken will return a random sha256 hash