+++ install.go
@@ -1,4 +1,4 @@
-package command
+package gwf
 
 import (
 	"fmt"
@@ -9,14 +9,13 @@ import (
 	"path/filepath"
 	"runtime"
 	"syscall"
-
-	gwf "github.com/RobyFerro/go-web-framework"
 )
 
 // Install method will install Go-Web
 type Install struct {
 	Signature   string
 	Description string
+	Args        string
 }
 
 // Register this command
@@ -26,18 +25,20 @@ func (c *Install) Register() {
 }
 
 // Run this command
-func (c *Install) Run(kernel *gwf.HttpKernel, args string, console map[string]interface{}) {
+func (c *Install) Run() {
 	var _, filename, _, _ = runtime.Caller(0)
-	if err := dir(filepath.Join(path.Dir(filename), "../../"), args); err != nil {
-		gwf.ProcessError(err)
+	if err := dir(filepath.Join(path.Dir(filename), "../../"), c.Args); err != nil {
+		ProcessError(err)
 	}
 }
 
 // Dir copies a whole directory recursively
 func dir(src string, dst string) error {
-	var err error
-	var fds []os.FileInfo
-	var srcInfo os.FileInfo
+	var (
+		err     error
+		fds     []os.FileInfo
+		srcInfo os.FileInfo
+	)
 
 	if srcInfo, err = os.Stat(src); err != nil {
 		return err
@@ -69,31 +70,34 @@ func dir(src string, dst string) error {
 
 // File copies a single file from src to dst
 func file(src, dst string) error {
-	var err error
-	var srcfd *os.File
-	var dstfd *os.File
-	var srcinfo os.FileInfo
+	var (
+		err      error
+		srcFile  *os.File
+		destFile *os.File
+		srcInfo  os.FileInfo
+	)
 
-	if srcfd, err = os.Open(src); err != nil {
+	if srcFile, err = os.Open(src); err != nil {
 		return err
 	}
-	defer srcfd.Close()
+	defer srcFile.Close()
 
-	if dstfd, err = os.Create(dst); err != nil {
+	if destFile, err = os.Create(dst); err != nil {
 		return err
 	}
-	defer dstfd.Close()
+	defer destFile.Close()
 
-	if _, err = io.Copy(dstfd, srcfd); err != nil {
+	if _, err = io.Copy(destFile, srcFile); err != nil {
 		return err
 	}
-	if srcinfo, err = os.Stat(src); err != nil {
+
+	if srcInfo, err = os.Stat(src); err != nil {
 		return err
 	}
 
 	if err := os.Chown(dst, syscall.Getuid(), syscall.Getgid()); err != nil {
-		gwf.ProcessError(err)
+		ProcessError(err)
 	}
 
-	return os.Chmod(dst, srcinfo.Mode())
+	return os.Chmod(dst, srcInfo.Mode())
 }