+++ jobs_create.go
@@ -1,8 +1,7 @@
-package command
+package gwf
 
 import (
 	"fmt"
-	"github.com/RobyFerro/go-web-framework"
 	"io/ioutil"
 	"path"
 	"path/filepath"
@@ -14,6 +13,7 @@ import (
 type JobCreate struct {
 	Signature   string
 	Description string
+	Args        string
 }
 
 // Register this command
@@ -22,22 +22,23 @@ func (c *JobCreate) Register() {
 	c.Description = "Create new async job" // Change command description
 }
 
-// Run this commmand
-func (c *JobCreate) Run(kernel *gwf.HttpKernel, args string, console map[string]interface{}) {
+// Run this command
+func (c *JobCreate) Run() {
 	var _, filename, _, _ = runtime.Caller(0)
 
-	splitName := strings.Split(strings.ToLower(args), "_")
+	splitName := strings.Split(strings.ToLower(c.Args), "_")
 	for i, name := range splitName {
 		splitName[i] = strings.Title(name)
 	}
 
 	cName := strings.Join(splitName, "")
-	input, _ := ioutil.ReadFile(filepath.Join(path.Dir(filename), "../../raw/job.raw"))
+	input, _ := ioutil.ReadFile(filepath.Join(path.Dir(filename), "raw/job.raw"))
 
 	cContent := strings.ReplaceAll(string(input), "@@TMP@@", cName)
-	cFile := fmt.Sprintf("%s/%s.go", gwf.GetDynamicPath("job"), strings.ToLower(args))
+	cFile := fmt.Sprintf("%s/%s.go", GetDynamicPath("job"), strings.ToLower(c.Args))
+
 	if err := ioutil.WriteFile(cFile, []byte(cContent), 0755); err != nil {
-		gwf.ProcessError(err)
+		ProcessError(err)
 	}
 
 	fmt.Printf("\nSUCCESS: Your %s job has been created at %s\n", cName, cFile)