+++ middleware_create.go
@@ -1,4 +1,4 @@
-package command
+package gwf
 
 import (
 	"fmt"
@@ -7,14 +7,13 @@ import (
 	"path/filepath"
 	"runtime"
 	"strings"
-
-	gwf "github.com/RobyFerro/go-web-framework"
 )
 
 // MiddlewareCreate will create a new http middleware
 type MiddlewareCreate struct {
 	Signature   string
 	Description string
+	Args        string
 }
 
 // Register this command
@@ -24,20 +23,22 @@ func (c *MiddlewareCreate) Register() {
 }
 
 // Run this command
-func (c *MiddlewareCreate) Run(kernel *gwf.HttpKernel, args string, console map[string]interface{}) {
+func (c *MiddlewareCreate) Run() {
 	var _, filename, _, _ = runtime.Caller(0)
-	splitName := strings.Split(strings.ToLower(args), "_")
+
+	splitName := strings.Split(strings.ToLower(c.Args), "_")
 	for i, name := range splitName {
 		splitName[i] = strings.Title(name)
 	}
 
 	cName := strings.Join(splitName, "")
-	input, _ := ioutil.ReadFile(filepath.Join(path.Dir(filename), "../../raw/middleware.raw"))
+	input, _ := ioutil.ReadFile(filepath.Join(path.Dir(filename), "raw/middleware.raw"))
 
 	cContent := strings.ReplaceAll(string(input), "@@TMP@@", cName)
-	cFile := fmt.Sprintf("%s/%s.go", gwf.GetDynamicPath("app/http/middleware"), strings.ToLower(args))
+	cFile := fmt.Sprintf("%s/%s.go", GetDynamicPath("app/http/middleware"), strings.ToLower(c.Args))
+
 	if err := ioutil.WriteFile(cFile, []byte(cContent), 0755); err != nil {
-		gwf.ProcessError(err)
+		ProcessError(err)
 	}
 
 	fmt.Printf("\nSUCCESS: Your %s middleware has been created at %s\n", cName, cFile)