+++ migration_create.go
@@ -1,11 +1,9 @@
-package command
+package gwf
 
 import (
 	"fmt"
 	"io/ioutil"
 	"time"
-
-	gwf "github.com/RobyFerro/go-web-framework"
 )
 
 // MigrationCreate will create a new migration
@@ -15,6 +13,7 @@ import (
 type MigrationCreate struct {
 	Signature   string
 	Description string
+	Args        string
 }
 
 // Register this command
@@ -24,24 +23,24 @@ func (c *MigrationCreate) Register() {
 }
 
 // Run this command
-func (c *MigrationCreate) Run(kernel *gwf.HttpKernel, args string, console map[string]interface{}) {
+func (c *MigrationCreate) Run() {
 	date := time.Now().Unix()
-	path := gwf.GetDynamicPath("database/migration")
+	path := GetDynamicPath("database/migration")
 
-	filenameUp := fmt.Sprintf("%s/%d_%s.up.sql", path, date, args)
-	filenameDown := fmt.Sprintf("%s/%d_%s.down.sql", path, date, args)
+	filenameUp := fmt.Sprintf("%s/%d_%s.up.sql", path, date, c.Args)
+	filenameDown := fmt.Sprintf("%s/%d_%s.down.sql", path, date, c.Args)
 
 	fmt.Printf("\nCreating new '%s'...\n", filenameUp)
 
 	if err := ioutil.WriteFile(filenameUp, []byte("/* MIGRATION UP */"), 0755); err != nil {
-		gwf.ProcessError(err)
+		ProcessError(err)
 	}
 
 	fmt.Printf("Created new up migration: %s\n", filenameUp)
 	fmt.Printf("Creating new '%s'...\n", filenameDown)
 
 	if err := ioutil.WriteFile(filenameDown, []byte("/* MIGRATION DOWN */"), 0755); err != nil {
-		gwf.ProcessError(err)
+		ProcessError(err)
 	}
 
 	fmt.Printf("Created new down migration: %s\n", filenameDown)