+++ migration_rollback.go
@@ -1,4 +1,4 @@
-package command
+package gwf
 
 import (
 	"fmt"
@@ -6,7 +6,6 @@ import (
 	"strconv"
 	"strings"
 
-	gwf "github.com/RobyFerro/go-web-framework"
 	"github.com/jinzhu/gorm"
 )
 
@@ -14,6 +13,7 @@ import (
 type MigrateRollback struct {
 	Signature   string
 	Description string
+	Args        string
 }
 
 // Register this command
@@ -23,21 +23,14 @@ func (c *MigrateRollback) Register() {
 }
 
 // Run this command
-func (c *MigrateRollback) Run(kernel *gwf.HttpKernel, args string, console map[string]interface{}) {
-	var db *gorm.DB
-	if err := kernel.Container.Invoke(func(client *gorm.DB) {
-		db = client
-	}); err != nil {
-		gwf.ProcessError(err)
-	}
-
-	step, _ := strconv.Atoi(args)
+func (c *MigrateRollback) Run(db *gorm.DB) {
+	step, _ := strconv.Atoi(c.Args)
 	batch := getLastBatch(db)
 
 	for i := 0; i < step; i++ {
 		var migrations []migration
 		if err := db.Order("created_at", true).Where("batch LIKE ?", batch).Find(&migrations).Error; err != nil {
-			gwf.ProcessError(err)
+			ProcessError(err)
 		}
 
 		// Execute given rollback
@@ -54,13 +47,13 @@ func rollbackMigrations(migrations []migration, db *gorm.DB) {
 		fmt.Printf("\nRolling back '%s' migration...\n", rollbackFile)
 
 		if payload, err := ioutil.ReadFile(rollbackFile); err != nil {
-			gwf.ProcessError(err)
+			ProcessError(err)
 		} else {
 			db.Exec(string(payload)).Row()
 		}
 
 		if err := db.Unscoped().Delete(&m).Error; err != nil {
-			gwf.ProcessError(err)
+			ProcessError(err)
 		}
 
 		fmt.Printf("Success! %s has been rolled back!", rollbackFile)