+++ model_create.go
@@ -1,4 +1,4 @@
-package command
+package gwf
 
 import (
 	"fmt"
@@ -7,14 +7,13 @@ import (
 	"path/filepath"
 	"runtime"
 	"strings"
-
-	gwf "github.com/RobyFerro/go-web-framework"
 )
 
 // ModelCreate will create a new Gorm model
 type ModelCreate struct {
 	Signature   string
 	Description string
+	Args        string
 }
 
 // Register this command
@@ -24,16 +23,17 @@ func (c *ModelCreate) Register() {
 }
 
 // Run this command
-func (c *ModelCreate) Run(kernel *gwf.HttpKernel, args string, console map[string]interface{}) {
+func (c *ModelCreate) Run() {
 	var _, filename, _, _ = runtime.Caller(0)
 
-	cName := strings.Title(strings.ToLower(args))
-	input, _ := ioutil.ReadFile(filepath.Join(path.Dir(filename), "../../raw/model.raw"))
+	cName := strings.Title(strings.ToLower(c.Args))
+	input, _ := ioutil.ReadFile(filepath.Join(path.Dir(filename), "raw/model.raw"))
 
 	cContent := strings.ReplaceAll(string(input), "@@TMP@@", cName)
-	cFile := fmt.Sprintf("%s/%s.go", gwf.GetDynamicPath("database/model"), strings.ToLower(args))
+	cFile := fmt.Sprintf("%s/%s.go", GetDynamicPath("database/model"), strings.ToLower(c.Args))
+
 	if err := ioutil.WriteFile(cFile, []byte(cContent), 0755); err != nil {
-		gwf.ProcessError(err)
+		ProcessError(err)
 	}
 
 	fmt.Printf("\nSUCCESS: Your model %s has been created at %s", cName, cFile)