+++ CHANGELOG.md
@@ -3,8 +3,16 @@
 All notable changes to the "Go-Web Framework" will be documented in this file.
 
 ## [Unreleased]
+
+## [v0.3.0-beta] - 2020-02-24
+### Changed
+- Basic type structure: controllers, services, models, and middleware.
+
+### Removed
+- Removed external configuration.
+
 ### Fixed
-- Add json tags in Auth (JWT) structure 
+- Fix authentication flow in JWTAuth (other fix)
 
 ## [v0.2.2-beta] - 2020-02-19
 ### Fixed